# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Openpage API",
        "version": "1.0.0"
    },
    "basePath": "/openpage-backend/public",
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Returns list of users",
                "description": "Returns users data",
                "operationId": "user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Search parameter or key word to search",
                        "type": "string"
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "this is offset or skip the records",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to be retrieved ",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/{uid}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Returns users data",
                "description": "Returns users data",
                "operationId": "user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "ID of the user that needs to be displayed",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid user id"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User registration into the system",
                "description": "User registration into the system",
                "operationId": "register",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name of user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name of user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address of user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User secured password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                }
            }
        },
        "/activate": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Activates user into the system",
                "description": "Activates user into the system",
                "operationId": "activate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Activation key",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
...
}